refactor search group creation
