+++ packages/itwin/grouping-mapping-widget/src/widget/components/extension/SearchExtension.scss
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------------------------------
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
+@import "~@itwin/itwinui-css/scss/variables";
+
+.gmw-search-form {
+  display: flex;
+  flex-direction: column;
+  gap: $iui-baseline;
+  margin-top: $iui-baseline;
+  > .gmw-search-actions {
+    display: flex;
+    justify-content: flex-end;
+    gap: $iui-s;
+  }
+}
